requirements.txt
pandas 
numpy
scikit-learn
matplotlib
seaborn
mysql-connector-python

-e .